+++ examples/datasets_and_pipelines/custom_dataset.py
@@ -40,7 +40,7 @@ Datasets work best in combination with `Pipelines` and are further compatible wi
 # Then you can filter the dataset first and load the data once you know which data-points you want to access.
 # We will discuss this later in the example.
 from itertools import product
-from typing import List, Optional, Union
+from typing import Optional, Union
 
 import pandas as pd
 
@@ -329,7 +329,7 @@ class CustomDatasetWithConfig(Dataset):
         data_folder: str,
         custom_config_para: bool = False,
         *,
-        groupby_cols: Optional[Union[List[str], str]] = None,
+        groupby_cols: Optional[Union[list[str], str]] = None,
         subset_index: Optional[pd.DataFrame] = None,
     ) -> None:
         self.data_folder = data_folder