+++ gaitmap/trajectory_reconstruction/trajectory_methods/_rts_kalman.py
@@ -13,11 +13,12 @@ from gaitmap.trajectory_reconstruction.trajectory_methods._kalman_numba_funcs im
     ForwardPassDependencies,
     SimpleZuptParameter,
     default_rts_kalman_forward_pass,
+    madgwick_mag_motion_update,
     madgwick_motion_update,
     rts_kalman_update_series,
     simple_navigation_equations,
 )
-from gaitmap.utils.consts import GF_POS, GF_VEL, SF_ACC, SF_GYR
+from gaitmap.utils.consts import GF_POS, GF_VEL, SF_ACC, SF_GYR, SF_MAG
 from gaitmap.utils.datatype_helper import SingleSensorData, SingleSensorStrideList, is_single_sensor_data
 from gaitmap.zupt_detection import NormZuptDetector
 
@@ -82,6 +83,9 @@ class RtsKalman(BaseTrajectoryMethod):
         They can be used as a measure of how good the filter worked and how accurate the results are.
     zupts_
         2D array indicating the start and the end samples of the detected ZUPTs for debug porpuses.
+    rotated_data_
+        The rotated data after applying the estimated orientation to the data.
+        The first sample of the data remain unrotated (initial orientation).
 
     Other Parameters
     ----------------
@@ -237,14 +241,14 @@ class RtsKalman(BaseTrajectoryMethod):
         zupts = zupt_detector.per_sample_zupts_
         self.zupts_ = zupt_detector.zupts_
 
-        gyro_data = np.deg2rad(data[SF_GYR].to_numpy())
-        acc_data = data[SF_ACC].to_numpy()
+        acc_data, gyro_data, mag_data = self._prepare_data()
 
         parameters = SimpleZuptParameter(level_walking=self.level_walking)
 
         states, covariances = rts_kalman_update_series(
             acc_data,
             gyro_data,
+            mag_data,
             initial_orientation,
             sampling_rate_hz,
             meas_noise,
@@ -283,6 +287,11 @@ class RtsKalman(BaseTrajectoryMethod):
         This should be overwritten by subclasses, that need custom internal functions/parameters.
         """
         return ForwardPassDependencies(motion_update_func=simple_navigation_equations, motion_update_func_parameters=())
+
+    def _prepare_data(self) -> tuple[np.ndarray, np.ndarray, Optional[np.ndarray]]:
+        gyro_data = np.deg2rad(self.data[SF_GYR].to_numpy())
+        acc_data = self.data[SF_ACC].to_numpy()
+        return acc_data, gyro_data, None
 
 
 class MadgwickRtsKalman(RtsKalman):
@@ -325,12 +334,17 @@ class MadgwickRtsKalman(RtsKalman):
         A high value should only be used if the sensor is moved slowly.
         A value of 0 is identical to just the Gyro Integration (i.e. identical to the
         :class:`~gaitmap.trajectory_reconstruction.RtsKalman`).
+    use_magnetometer
+        Flag to control if the magnetometer should be used in the Madgwick filter.
+        Note, that the rest of the algorithm does not change based on this parameter.
+        The Kalman update steps and the error propagation still only consider the accelerometer and gyroscope data.
+        If True, the data is expected to have the `mag_x, mag_y, mag_z` columns.
 
     Attributes
     ----------
     orientation_
         The rotations as a *SingleSensorOrientationList*, including the initial orientation.
-        This means the there are len(data) + 1 orientations.
+        This means there are len(data) + 1 orientations.
     orientation_object_
         The orientations as a single scipy Rotation object
     position_
@@ -342,6 +356,9 @@ class MadgwickRtsKalman(RtsKalman):
         They can be used as a measure of how good the filter worked and how accurate the results are.
     zupts_
         2D array indicating the start and the end samples of the detected ZUPTs for debug purposes.
+    rotated_data_
+        The rotated data after applying the estimated orientation to the data.
+        The first sample of the data remain unrotated (initial orientation).
 
     Other Parameters
     ----------------
@@ -358,6 +375,7 @@ class MadgwickRtsKalman(RtsKalman):
     """
 
     madgwick_beta: float
+    use_magnetometer: bool
 
     def __init__(
         self,
@@ -374,8 +392,10 @@ class MadgwickRtsKalman(RtsKalman):
             )
         ),
         madgwick_beta: float = 0.2,
+        use_magnetometer: bool = False,
     ) -> None:
         self.madgwick_beta = madgwick_beta
+        self.use_magnetometer = use_magnetometer
         super().__init__(
             initial_orientation=initial_orientation,
             zupt_variance=zupt_variance,
@@ -387,6 +407,17 @@ class MadgwickRtsKalman(RtsKalman):
         )
 
     def _prepare_forward_pass_dependencies(self) -> ForwardPassDependencies:
+        # This is kind of the wrong place for this check, but it is simpler then completely changing the structure
         return ForwardPassDependencies(
-            motion_update_func=madgwick_motion_update, motion_update_func_parameters=(self.madgwick_beta,)
+            motion_update_func=madgwick_mag_motion_update if self.use_magnetometer else madgwick_motion_update,
+            motion_update_func_parameters=(self.madgwick_beta,),
         )
+
+    def _prepare_data(self) -> tuple[np.ndarray, np.ndarray, Optional[np.ndarray]]:
+        gyro_data = np.deg2rad(self.data[SF_GYR].to_numpy())
+        acc_data = self.data[SF_ACC].to_numpy()
+        if self.use_magnetometer is False:
+            return acc_data, gyro_data, None
+        is_single_sensor_data(self.data, frame="sensor", raise_exception=True, check_mag=True)
+        mag_data = self.data[SF_MAG].to_numpy()
+        return acc_data, gyro_data, mag_data