+++ tests/conftest.py
@@ -1,11 +1,11 @@
 import random
-from typing import Any, Dict
+from typing import Any
 
 import numpy as np
 import pandas as pd
 import pytest
 from numpy.testing import assert_almost_equal, assert_array_equal
-from pandas._testing import assert_frame_equal, assert_series_equal
+from pandas.testing import assert_frame_equal, assert_series_equal
 from scipy.spatial.transform import Rotation
 from tpcp import BaseTpcpObject
 
@@ -20,7 +20,6 @@ from gaitmap.example_data import (
     get_healthy_example_stride_events,
     get_ms_example_imu_data,
 )
-from tests._regression_utils import PyTestSnapshotTest
 
 try:
     from pomegranate import GeneralMixtureModel, State
@@ -35,19 +34,6 @@ def reset_random_seed() -> None:
     random.seed(10)
 
 
-@pytest.fixture()
-def snapshot(request):
-    with PyTestSnapshotTest(request) as snapshot_test:
-        yield snapshot_test
-
-
-def pytest_addoption(parser) -> None:
-    group = parser.getgroup("snapshottest")
-    group.addoption(
-        "--snapshot-update", action="store_true", default=False, dest="snapshot_update", help="Update the snapshots."
-    )
-
-
 healthy_example_imu_data = pytest.fixture()(get_healthy_example_imu_data)
 healthy_example_imu_data_ic_stride = pytest.fixture()(get_healthy_example_imu_data_ic_stride)
 ms_example_imu_data = pytest.fixture()(get_ms_example_imu_data)
@@ -59,7 +45,7 @@ healthy_example_orientation = pytest.fixture()(get_healthy_example_orientation)
 healthy_example_position = pytest.fixture()(get_healthy_example_position)
 
 
-def _get_params_without_nested_class(instance: BaseTpcpObject) -> Dict[str, Any]:
+def _get_params_without_nested_class(instance: BaseTpcpObject) -> dict[str, Any]:
     return {k: v for k, v in instance.get_params().items() if not hasattr(v, "get_params")}